+++ src/overload_numpy/dispatch.py
@@ -7,8 +7,8 @@ from __future__ import annotations
 from functools import singledispatch
 from typing import TYPE_CHECKING, Any, NoReturn, final
 
-# LOCALFOLDER
-from overload_numpy.npinfo import _NOT_DISPATCHED, _NumPyInfo
+# LOCAL
+from overload_numpy.npinfo import _NOT_DISPATCHED, _NumPyFuncOverloadInfo
 
 if TYPE_CHECKING:
     # STDLIB
@@ -22,38 +22,38 @@ __all__: list[str] = []
 ##############################################################################
 
 
-# A stand-in function for a not-dispatched `_NumPyInfo`. If
-# `_NumPyInfo.validate_types` is used in `__array_function__` then this function
+# A stand-in function for a not-dispatched `_NumPyFuncOverloadInfo`. If
+# `_NumPyFuncOverloadInfo.validate_types` is used in `__array_function__` then this function
 # will never be called.
 def _notdispatched(*args: Any, **kwarg: Any) -> NoReturn:
     raise NotImplementedError("not dispatched")
 
 
-_notdispatched_info = _NumPyInfo(
+_notdispatched_info = _NumPyFuncOverloadInfo(
     func=_notdispatched, implements=_notdispatched, types=_NOT_DISPATCHED, dispatch_on=object
 )
-# The not-dispatched `_NumPyInfo`. All `_Dispatcher` start with this as the base
-# `_NumPyInfo`.
+# The not-dispatched `_NumPyFuncOverloadInfo`. All `_Dispatcher` start with this as the base
+# `_NumPyFuncOverloadInfo`.
 
 
 @final
 class _Dispatcher:
-    """Single dispatch."""
+    """`~functools.singledispatch` instance."""
 
     def __init__(self) -> None:
-        # Make default NotImplementedError _NumPyInfo
-        # This return value is necessary until the functional form of
-        # singledispatch works with MyPy for adding different return types.
-        # Until then this is restricted to the one return type.
-        # This cannot be built in mypyc v0.971
+        # Make default NotImplementedError
+        # `~overload_numpy.npinfo._NumPyFuncOverloadInfo` This return value is
+        # necessary until the functional form of singledispatch works with MyPy
+        # for adding different return types. Until then this is restricted to
+        # the one return type. This cannot be built in mypyc v0.971
         @singledispatch
-        def dispatcher(obj: object) -> _NumPyInfo:
+        def dispatcher(obj: object, /) -> _NumPyFuncOverloadInfo:
             return _notdispatched_info
 
-        self._dispatcher: functools._SingleDispatchCallable[_NumPyInfo]
+        self._dispatcher: functools._SingleDispatchCallable[_NumPyFuncOverloadInfo]
         self._dispatcher = dispatcher
 
-    def __call__(self, obj: object, /) -> _NumPyInfo:
-        """Get correct _NumPyInfo for the calling object's type."""
-        npinfo: _NumPyInfo = self._dispatcher(obj)
+    def __call__(self, obj: object, /) -> _NumPyFuncOverloadInfo:
+        """Get correct `~overload_numpy.npinfo._NumPyFuncOverloadInfo` for the calling object's type."""
+        npinfo: _NumPyFuncOverloadInfo = self._dispatcher(obj)
         return npinfo