+++ tests/unit/test_npinfo.py
@@ -4,6 +4,8 @@
 from __future__ import annotations
 
 # STDLIB
+import pickle
+from copy import copy, deepcopy
 from numbers import Complex, Rational, Real
 from typing import TYPE_CHECKING, Callable
 
@@ -11,14 +13,31 @@ from typing import TYPE_CHECKING, Callable
 import numpy as np
 import pytest
 
-# LOCALFOLDER
+# LOCAL
 from .test_constraints import TypeConstraint_TestBase
 from overload_numpy.constraints import Covariant, Invariant, TypeConstraint
-from overload_numpy.npinfo import _NOT_DISPATCHED, _NotDispatched, _NumPyInfo
+from overload_numpy.npinfo import (
+    _NOT_DISPATCHED,
+    _NotDispatched,
+    _NumPyFuncOverloadInfo,
+)
 
 if TYPE_CHECKING:
     # STDLIB
     from types import FunctionType
+
+
+##############################################################################
+
+
+class Magnitude:
+    def __init__(self, x) -> None:
+        self._x = x
+
+
+def add(obj1, obj2):
+    return obj1._x + obj2
+
 
 ##############################################################################
 # TESTS
@@ -32,57 +51,59 @@ class Test__NotDispatched(TypeConstraint_TestBase):
 
     @pytest.fixture(scope="class")
     def constraint(self, constraint_cls) -> TypeConstraint:
-        return _NOT_DISPATCHED
+        return constraint_cls()
+
+    # ===============================================================
+    # Method Tests
 
     def test_validate_type(self, constraint) -> None:
         assert constraint.validate_type(None) is False
+
+    # ===============================================================
+    # Usage Tests
+
+    def test_instance(self, constraint_cls) -> None:
+        assert isinstance(_NOT_DISPATCHED, constraint_cls)
 
 
 ##############################################################################
 
 
-class Test__NumPyInfo:
-    """Test :class:`overload_numpy.dispatch._NumPyInfo`."""
+class Test__NumPyFuncOverloadInfo:
+    """Test :class:`overload_numpy.dispatch._NumPyFuncOverloadInfo`."""
 
     @pytest.fixture(scope="class")
     def custom_cls(self):
-        class Magnitude:
-            def __init__(self, x) -> None:
-                self._x = x
-
         return Magnitude
 
     @pytest.fixture(scope="class")
     def implements_info(self) -> tuple[Callable, Callable]:
-        def add(obj1, obj2):
-            return obj1._x + obj2
-
         return add, np.add
 
     @pytest.fixture
-    def npinfo(self, custom_cls: type, implements_info: tuple[FunctionType, Callable]) -> _NumPyInfo:
+    def npinfo(self, custom_cls: type, implements_info: tuple[FunctionType, Callable]) -> _NumPyFuncOverloadInfo:
         # TypeConstraint
         tinfo = (Covariant(custom_cls), Covariant(Real))
 
-        return _NumPyInfo(*implements_info, tinfo, dispatch_on=object)
+        return _NumPyFuncOverloadInfo(*implements_info, tinfo, dispatch_on=object)
 
     # ===============================================================
 
     def test_init_error_func(self):
         with pytest.raises(TypeError, match="func must be callable"):
-            _NumPyInfo(func=None, implements=np.add, types=Invariant(Real), dispatch_on=object)
+            _NumPyFuncOverloadInfo(func=None, implements=np.add, types=Invariant(Real), dispatch_on=object)
 
     def test_init_error_implements(self):
         with pytest.raises(TypeError, match="implements must be callable"):
-            _NumPyInfo(func=lambda x: x, implements=None, types=Invariant(Real), dispatch_on=object)
+            _NumPyFuncOverloadInfo(func=lambda x: x, implements=None, types=Invariant(Real), dispatch_on=object)
 
     def test_init_error_types(self):
         with pytest.raises(TypeError, match="types"):
-            _NumPyInfo(func=lambda x: x, implements=np.add, types=None, dispatch_on=object)
+            _NumPyFuncOverloadInfo(func=lambda x: x, implements=np.add, types=None, dispatch_on=object)
 
     def test_init_error_dispatch_on(self):
         with pytest.raises(TypeError, match="dispatch"):
-            _NumPyInfo(func=lambda x: x, implements=np.add, types=Invariant(Real), dispatch_on=None)
+            _NumPyFuncOverloadInfo(func=lambda x: x, implements=np.add, types=Invariant(Real), dispatch_on=None)
 
     def test_init(self, npinfo):
         # The pytest fixture proves it passes.
@@ -92,18 +113,18 @@ class Test__NumPyInfo:
 
     @pytest.mark.xfail  # limited by mypyc
     def test_validate_types_NotImplemented(self, implements_info):
-        npinfo = _NumPyInfo(*implements_info, types=NotImplemented, dispatch_on=object)
+        npinfo = _NumPyFuncOverloadInfo(*implements_info, types=NotImplemented, dispatch_on=object)
 
         assert npinfo.validate_types(()) is False
 
     def test_validate_types_NotDispatched(self, implements_info):
-        npinfo = _NumPyInfo(*implements_info, types=_NOT_DISPATCHED, dispatch_on=object)
+        npinfo = _NumPyFuncOverloadInfo(*implements_info, types=_NOT_DISPATCHED, dispatch_on=object)
 
         assert npinfo.validate_types(()) is False
 
     def test_validate_types_TypeConstraint(self, implements_info):
         # TODO! go through more TypeConstraint
-        npinfo = _NumPyInfo(*implements_info, types=Invariant(Real), dispatch_on=object)
+        npinfo = _NumPyFuncOverloadInfo(*implements_info, types=Invariant(Real), dispatch_on=object)
 
         assert npinfo.validate_types((Real,)) is True
         assert npinfo.validate_types((Real, Real)) is True
@@ -118,7 +139,9 @@ class Test__NumPyInfo:
 
     def test_validate_types_Collection(self, implements_info):
         # TODO! go through more TypeConstraint
-        npinfo = _NumPyInfo(*implements_info, types=(Invariant(Real), Invariant(Complex)), dispatch_on=object)
+        npinfo = _NumPyFuncOverloadInfo(
+            *implements_info, types=(Invariant(Real), Invariant(Complex)), dispatch_on=object
+        )
 
         assert npinfo.validate_types((Real,)) is True
         assert npinfo.validate_types((Real, Real)) is True
@@ -130,3 +153,16 @@ class Test__NumPyInfo:
         assert npinfo.validate_types((Complex,)) is True
         assert npinfo.validate_types((Complex, Real)) is True
         assert npinfo.validate_types((Complex, Rational)) is False
+
+    # ===============================================================
+    # Usage Tests
+
+    @pytest.mark.incompatible_with_mypyc
+    def test_serialization(self, implements_info) -> None:
+        # copying
+        assert copy(implements_info) == implements_info
+        assert deepcopy(implements_info) == implements_info
+
+        # pickling
+        dumps = pickle.dumps(implements_info)
+        assert pickle.loads(dumps) == implements_info