+++ CHANGES.rst
@@ -4,7 +4,9 @@ Version 0.1
 New Features
 ------------
 
-- ``dispatch_on`` type information is added to the ``_NumPyInfo`` object to
+- ``_NumPyInfo`` is renamed to ``_NumPyFuncOverloadInfo`` [#16]
+
+- ``dispatch_on`` type information is added to the ``_NumPyFuncOverloadInfo`` object to
   which it dispaches. [#21]
 
 - Arguments to ``TypeConstraint.validate_types`` (and subclasses) are now